+++ ion/integration/ais/common/metadata_cache.py
@@ -25,11 +25,12 @@ from ion.services.coi.resource_registry.association_client import AssociationCli
 
 from ion.services.dm.inventory.association_service import AssociationServiceClient, AssociationServiceError, ASSOCIATION_GET_STAR_MSG_TYPE
 from ion.services.dm.inventory.association_service import PREDICATE_OBJECT_QUERY_TYPE, \
-    SUBJECT_PREDICATE_QUERY_TYPE, IDREF_TYPE
+    SUBJECT_PREDICATE_QUERY_TYPE, IDREF_TYPE, PREDICATE_REFERENCE_TYPE
 from ion.services.coi.datastore_bootstrap.ion_preload_config import TYPE_OF_ID, \
     DATASET_RESOURCE_TYPE_ID, DATASOURCE_RESOURCE_TYPE_ID, HAS_A_ID, OWNED_BY_ID
 
-PREDICATE_REFERENCE_TYPE = object_utils.create_type_identifier(object_id=25, version=1)
+from ion.integration.ais.common.ais_utils import AIS_Mixin
+
 
 #
 # Common Metadata Constants
@@ -76,7 +77,7 @@ UPDATE_INTERVAL_SECONDS = 'update_interval_seconds'
 VISUALIZATION_URL = 'visualization_url'
 VISIBILITY = 'visibility'
 
-class MetadataCache(object):
+class MetadataCache(AIS_Mixin):
     
     __metadata = {}
 
@@ -129,7 +130,7 @@ class MetadataCache(object):
         log.debug('loadDataSets()')
 
         # Get the list of dataset resource IDs
-        dSetResults = yield self.__findResourcesOfType(DATASET_RESOURCE_TYPE_ID)
+        dSetResults = yield self.findResourcesOfType(DATASET_RESOURCE_TYPE_ID)
         if dSetResults == None:
             log.error('Error finding dataset resources.')
             defer.returnValue(False)
@@ -162,7 +163,7 @@ class MetadataCache(object):
         log.debug('loadDataSources()')
         
         # Get the list of datasource resource IDs
-        dSourceResults = yield self.__findResourcesOfType(DATASOURCE_RESOURCE_TYPE_ID)
+        dSourceResults = yield self.findResourcesOfType(DATASOURCE_RESOURCE_TYPE_ID)
         if dSourceResults == None:
             log.error('Error finding datasource resources.')
             defer.returnValue(False)
@@ -422,87 +423,6 @@ class MetadataCache(object):
         defer.returnValue(returnValue)
 
 
-    @defer.inlineCallbacks
-    def getAssociatedSource(self, dSetID):
-        """
-        Worker class private method to get the data source that associated
-        with a given data set.  
-        """
-
-        if dSetID is None:
-            log.error('getAssociatedSource: dSetID is None')
-            defer.returnValue(None)
-
-        log.debug('getAssociatedSource for dSetID %s' %(dSetID))
-
-        qmsg = yield self.mc.create_instance(PREDICATE_OBJECT_QUERY_TYPE)
-        pair = qmsg.pairs.add()
-        pair.object = qmsg.CreateObject(IDREF_TYPE)
-        pair.object.key = dSetID
-        pair.predicate = qmsg.CreateObject(PREDICATE_REFERENCE_TYPE)
-        pair.predicate.key = HAS_A_ID
-
-        pair = qmsg.pairs.add()
-        pair.object = qmsg.CreateObject(IDREF_TYPE)
-        pair.object.key = DATASOURCE_RESOURCE_TYPE_ID
-        pair.predicate = qmsg.CreateObject(PREDICATE_REFERENCE_TYPE)
-        pair.predicate.key = TYPE_OF_ID
-        try:
-            results = yield self.asc.get_subjects(qmsg)
-        except:
-            log.exception('Error getting associated data source for Dataset: %s' % dSetID)
-            defer.returnValue(None)
-
-        dsrcs = [str(x.key) for x in results.idrefs]
-
-        # we expect one:
-        if len(dsrcs) != 1:
-            log.error('Expected 1 datasource, got %d' % len(dsrcs))
-            defer.returnValue(None)
-
-        log.debug('getAssociatedSource() exit: returning: %s' % dsrcs[0])
-
-        defer.returnValue(dsrcs[0])
-
-    @defer.inlineCallbacks
-    def getAssociatedDatasets(self, dSource):
-        """
-        Worker class private method to get the data sets that associated
-        with a given data source.  
-        """
-        log.debug('getAssociatedDatasets() entry')
-        
-        dSetList = []
-
-        if dSource is None:
-            log.error('getAssociatedDatasets: dSource parameter is None')
-            defer.returnValue(dSetList)
-
-        qmsg = yield self.mc.create_instance(ASSOCIATION_GET_STAR_MSG_TYPE)
-        pair = qmsg.subject_pairs.add()
-        pair.subject = qmsg.CreateObject(IDREF_TYPE)
-        pair.subject.key = dSource.ResourceIdentity
-        pair.predicate = qmsg.CreateObject(PREDICATE_REFERENCE_TYPE)
-        pair.predicate.key = HAS_A_ID
-
-        pair = qmsg.object_pairs.add()
-        pair.object = qmsg.CreateObject(IDREF_TYPE)
-        pair.object.key = DATASET_RESOURCE_TYPE_ID
-        pair.predicate = qmsg.CreateObject(PREDICATE_REFERENCE_TYPE)
-        pair.predicate.key = TYPE_OF_ID
-        try:
-            results = yield self.asc.get_star(qmsg)
-        except:
-            log.error('Error getting associated data sets for Datasource: ' + \
-                      dSource.ResourceIdentity)
-            defer.returnValue([])
-
-        dsets = [str(x.key) for x in results.idrefs]
-
-        log.info('Datasource %s has %d associated datasets.' %(dSource.ResourceIdentity, len(dsets)))
-        log.debug('getAssociatedDatasets) exit: returning: %s' %(str(dsets)))
-
-        defer.returnValue(dsets)
 
 
     @defer.inlineCallbacks
@@ -585,7 +505,7 @@ class MetadataCache(object):
             dSetMetadata[DSET] = dSet
             dSetMetadata[DSOURCE_ID] = yield self.getAssociatedSource(dSet.ResourceIdentity)
             dSetMetadata[RESOURCE_ID] = dSet.ResourceIdentity
-            dSetMetadata[OWNER_ID] = yield self.__getAssociatedOwner(dSet.ResourceIdentity)
+            dSetMetadata[OWNER_ID] = yield self.getAssociatedOwner(dSet.ResourceIdentity)
             for attrib in dSet.root_group.attributes:
                 #log.debug('Root Attribute: %s = %s'  % (str(attrib.name), str(attrib.GetValue())))
                 if attrib.name == TITLE:
@@ -682,93 +602,7 @@ class MetadataCache(object):
         else:
             log.info('data source %s is not ACTIVE: Not caching.' %(dSource.ResourceIdentity))
 
-                
 
-    @defer.inlineCallbacks
-    def __findResourcesOfType(self, resourceType):
-        """
-        A utility method to find all resources of the given type (resourceType).
-        """
-
-        request = yield self.mc.create_instance(PREDICATE_OBJECT_QUERY_TYPE)
-
-        #
-        # Set up a resource type search term using:
-        # - TYPE_OF_ID as predicate
-        # - object of type: resourceType parameter as object
-        #
-        pair = request.pairs.add()
-    
-        # ..(predicate)
-        pref = request.CreateObject(PREDICATE_REFERENCE_TYPE)
-        pref.key = TYPE_OF_ID
-
-        pair.predicate = pref
-
-        # ..(object)
-        type_ref = request.CreateObject(IDREF_TYPE)
-        type_ref.key = resourceType
-        
-        pair.object = type_ref
-
-        try:
-            result = yield self.asc.get_subjects(request)
-
-        except AssociationServiceError:
-            log.error('__findResourcesOfType: association error!')
-            defer.returnValue(None)
-
-        defer.returnValue(result)
-
-
-    @defer.inlineCallbacks
-    def __getAssociatedOwner(self, dsID):
-        """
-        Worker class method to find the owner associated with a data set.
-        This is a public method because it can be called from the
-        findDataResourceDetail worker class.
-        """
-        log.debug('getAssociatedOwner() entry')
-
-        request = yield self.mc.create_instance(SUBJECT_PREDICATE_QUERY_TYPE)
-
-        #
-        # Set up an owned_by_id search term using:
-        # - OWNED_BY_ID as predicate
-        # - LCS_REFERENCE_TYPE object set to ACTIVE as object
-        #
-        pair = request.pairs.add()
-
-        # ..(predicate)
-        pref = request.CreateObject(PREDICATE_REFERENCE_TYPE)
-        pref.key = OWNED_BY_ID
-
-        pair.predicate = pref
-
-        # ..(subject)
-        type_ref = request.CreateObject(IDREF_TYPE)
-        type_ref.key = dsID
-        
-        pair.subject = type_ref
-
-        log.info('Calling get_objects with dsID: ' + dsID)
-
-        try:
-            result = yield self.asc.get_objects(request)
-        
-        except AssociationServiceError:
-            log.error('getAssociatedOwner: association error!')
-            defer.returnValue(None)
-
-        if len(result.idrefs) == 0:
-            log.error('Owner not found!')
-            defer.returnValue('OWNER NOT FOUND!')
-        elif len(result.idrefs) == 1:
-            log.debug('getAssociatedOwner() exit')
-            defer.returnValue(result.idrefs[0].key)
-        else:
-            log.error('More than 1 owner found!')
-            defer.returnValue('MULTIPLE OWNERS!')
 
 
     def __printMetadata(self, resType, res):
